agentscore-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
py-algorand-sdk==2.11.1
python-dotenv==1.1.0
pydantic==2.12.5
fastmcp==2.12.5
x402==2.6.0
python-telegram-bot[webhooks]>=20.0
httpx>=0.27.0
cryptography>=43.0.0
anthropic>=0.40.0
